backend: OrdineController: modify POST method

+++ src/main/java/it/unical/inf/ea/trintedapp/data/entities/Utente.java
@@ -77,7 +77,7 @@ public class Utente {
     private List<Ordine> ordiniUtente;
 
     @Column(name = "saldo")
-    private Float saldo;
+    private Double saldo;
 
     @Column(name = "rating_genrale")
     private Float ratingGenerale;
